chore: organize tests

+++ src/card/suit_range.rs
@@ -15,16 +15,16 @@ pub struct SuitRange {
 impl SuitRange {
     pub fn new(start: Suit, end: Suit) -> Self {
         Self {
-            start: start.into(),
-            end: end.into(),
+            start: u8::from(start).into(),
+            end: u8::from(end).into(),
             inclusive: false,
         }
     }
 
     pub fn inclusive(start: Suit, end: Suit) -> Self {
         Self {
-            start: start.into(),
-            end: end.into(),
+            start: u8::from(start).into(),
+            end: u8::from(end).into(),
             inclusive: true,
         }
     }
